Add OSM in library
