+++ geojsonify/src/main/java/com/nextome/geojsonify/GeoJsonify.java
@@ -6,14 +6,20 @@ import android.net.Uri;
 import com.google.android.gms.maps.GoogleMap;
 import com.mapbox.mapboxsdk.maps.MapboxMap;
 
+import org.osmdroid.views.MapView;
+
 import java.util.List;
 
 public class GeoJsonify {
     public static void geoJsonifyMap(GoogleMap map, List<Uri> jsonUris, List<Integer> jsonColors, Context context){
-        JsonifyGoogleMaps.GeoJsonifyMap(map, jsonUris, jsonColors, context);
+        JsonifyGoogleMaps.geoJsonifyMap(map, jsonUris, jsonColors, context);
     }
 
     public static void geoJsonifyMap(MapboxMap map, List<Uri> jsonUris, List<Integer> jsonColors, Context context){
-        JsonifyMapBox.GeoJsonifyMap(map, jsonUris, jsonColors, context);
+        JsonifyMapBox.geoJsonifyMap(map, jsonUris, jsonColors, context);
+    }
+
+    public static void geoJsonifyMap(MapView map, List<Uri> jsonUris, List<Integer> jsonColors, Context context){
+        JsonifyOSM.geoJsonifyMap(map, jsonUris, jsonColors, context);
     }
 }
